+++ packages/core/src/app/address/PayPalAxo/index.ts
@@ -1,0 +1,3 @@
+export {default as PoweredByPaypalConnectLabel} from './PoweredByPaypalConnectLabel';
+export {default as usePayPalConnectAddress} from './usePayPalConnectAddress';
+export { isPayPalConnectAddress } from './utils';