+++ packages/core/src/app/address/StaticAddress.spec.tsx
@@ -13,7 +13,13 @@ import { getCountries } from '../geography/countries.mock';
 import { getAddress } from './address.mock';
 import AddressType from './AddressType';
 import { getAddressFormFields } from './formField.mock';
+import * as usePayPalConnectAddress from './PayPalAxo/usePayPalConnectAddress';
 import StaticAddress, { StaticAddressProps } from './StaticAddress';
+
+jest.mock('@bigcommerce/checkout/ui', () => ({
+    ...jest.requireActual('@bigcommerce/checkout/ui'),
+    IconPayPalConnectSmall: () => (<div data-test="pp-connect-icon">IconPayPalConnectSmall</div>)
+}));
 
 describe('StaticAddress Component', () => {
     let StaticAddressTest: FunctionComponent<StaticAddressProps>;
@@ -37,6 +43,12 @@ describe('StaticAddress Component', () => {
 
         jest.spyOn(checkoutState.data, 'getShippingAddressFields').mockReturnValue(
             getAddressFormFields(),
+        );
+
+        jest.spyOn(usePayPalConnectAddress, 'default').mockImplementation(
+            jest.fn().mockImplementation(() => ({
+                isPayPalConnectAddress: () => false,
+            }))
         );
 
         StaticAddressTest = (props) => (
@@ -146,5 +158,34 @@ describe('StaticAddress Component', () => {
         );
 
         expect(container.html()).not.toBe('');
+    });
+
+    describe('provider icon', () => {
+        it('should not show icon, address is not from PP Connect', () => {
+            const container = render(
+                <StaticAddressTest
+                    {...defaultProps}
+                />
+            );
+
+            expect(container.find('[data-test="pp-connect-icon"]')).toHaveLength(0);
+        });
+
+        it('should show icon for PP Connect address', () => {
+            jest.spyOn(usePayPalConnectAddress, 'default').mockImplementation(
+                jest.fn().mockImplementation(() => ({
+                    isPayPalAxoEnabled: true,
+                    paypalConnectAddresses: [defaultProps.address]
+                }))
+            );
+
+            const container = render(
+                <StaticAddressTest
+                    {...defaultProps}
+                />
+            );
+
+            expect(container.find('[data-test="pp-connect-icon"]')).toHaveLength(1);
+        });
     });
 });